requirements.txt
opencv-python==4.8.0.74
pandas==2.0.3
scipy==1.11.1
scikit-image==0.21.0
Pillow==10.0.0
matplotlib==3.7.2
pydicom==2.4.2
torch==2.0.1
torchvision==0.15.2
mlflow==2.7.1
tqdm==4.65.0
lpips==0.1.4
SimpleITK==2.3.0
tensorboardX==2.6.2.2
tensorflow==2.16.2
fastremap==1.14.0
lungmask==0.2.20
python-gdcm==3.0.24.1